requirements.txt
# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# OpenSearch
opensearch-py>=2.4.0

# Vector operations
numpy>=1.24.0
scikit-learn>=1.3.0

# Data processing
pandas>=2.0.0
pydantic>=2.0.0

# Web framework for API
fastapi>=0.104.0
uvicorn>=0.24.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Text processing
langchain>=0.1.0
langchain-community>=0.0.10
langchain-aws>=0.1.0

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 